build_regressor経由でPolyregressorを準備した
