+++ main.py
@@ -1,6 +1,7 @@
 import numpy as np
 from matplotlib.figure import Figure
 import japanize_matplotlib as _
+from regressor import build_regressor
 
 def calculate_score(y, y_pred, eps_score):
     norm_diff = np.sum(np.abs(y - y_pred))
@@ -35,25 +36,6 @@ def save_graph(
     ax.legend()
     fig.savefig(filename)
 
-class PolyRegressor:
-    def __init__(self, d):
-        self.d = d
-        self.p = np.arange(d+1)[np.newaxis, :]
-
-    def fit(self, x_sample, y_sample):
-         ## Xを作る
-        x_s = x_sample[:, np.newaxis]
-        X_s = x_s ** self.p
-        ##係数aを求める
-        y_s = y_sample[:, np.newaxis]
-        X_inv = np.linalg.inv(X_s.T @ X_s)
-        self.a =X_inv @ X_s.T @ y_s
-
-    def predict(self, x):
-        ## yの予測値を計算
-        y_pred = np.squeeze((x[:, np.newaxis] ** self.p) @ self.a)
-        return y_pred
-
 def main():
     # 実験条件
     x_min = -1
@@ -62,9 +44,14 @@ def main():
     n_test = 101
     noise_ratio = 0.05
     eps_score = 1e-8
-    # 多項式フィッティングの設定
-    d = 3
-    regressor = PolyRegressor(d)
+    # 回帰分析に関する設定
+    regressor_name = 'poly'
+    regressor_kwargs = dict(
+        poly = dict(
+            d = 3,
+        ),
+    )
+    regressor = build_regressor(regressor_name, regressor_kwargs)
     # x,f(x)の準備
     x = np.linspace(x_min, x_max,n_test)
     y = np.sin(np.pi * x)
